refactor: enable h2 console, add @Valid, package
