+++ src/main/java/com/ddd/moodof/adapter/infrastructure/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.ddd.moodof.infrastructure.auth.exception;
+package com.ddd.moodof.adapter.infrastructure.auth.exception;
 
 public class InvalidTokenException extends RuntimeException {
     public InvalidTokenException(String message) {
