+++ src/main/java/com/ddd/moodof/adapter/presentation/StoragePhotoController.java
@@ -1,4 +1,4 @@
-package com.ddd.moodof.presentation;
+package com.ddd.moodof.adapter.presentation;
 
 import com.ddd.moodof.application.StoragePhotoService;
 import com.ddd.moodof.application.dto.StoragePhotoDTO;
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
 import java.net.URI;
 
 @RequiredArgsConstructor
@@ -20,7 +21,7 @@ public class StoragePhotoController {
     private final StoragePhotoService storagePhotoService;
 
     @PostMapping
-    public ResponseEntity<StoragePhotoDTO.Response> create(@RequestBody StoragePhotoDTO.Create request, @LoginUserId Long userId) {
+    public ResponseEntity<StoragePhotoDTO.Response> create(@RequestBody @Valid StoragePhotoDTO.Create request, @LoginUserId Long userId) {
         StoragePhotoDTO.Response response = storagePhotoService.create(request, userId);
         return ResponseEntity.created(URI.create(API_STORAGE_PHOTO + "/" + response.getId())).body(response);
     }
