+++ src/main/java/com/ddd/moodof/domain/model/storage/photo/StoragePhotoRepository.java
@@ -1,4 +1,4 @@
-package com.ddd.moodof.domain.storage.photo;
+package com.ddd.moodof.domain.model.storage.photo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
