+++ src/main/java/com/ddd/moodof/domain/model/user/User.java
@@ -1,4 +1,4 @@
-package com.ddd.moodof.domain.entity;
+package com.ddd.moodof.domain.model.user;
 
 
 import lombok.Getter;
