+++ src/main/resources/application.yml
@@ -13,6 +13,9 @@ spring:
       hibernate:
         show_sql: true
         format_sql: true
+  h2:
+    console:
+      enabled: true
 logging.level:
   org.hibernate.SQL: debug
 # org.hibernate.type: trace
